+++ mayan/apps/task_manager/handlers.py
@@ -1,0 +1,5 @@
+from .utils import purge_periodic_tasks
+
+
+def handler_perform_upgrade(sender, **kwargs):
+    purge_periodic_tasks()