+++ mayan/apps/task_manager/tests/mixins.py
@@ -1,6 +1,26 @@
+from django_celery_beat.models import IntervalSchedule, PeriodicTask
+
 from ..classes import Worker, CeleryQueue
 
-from .literals import TEST_QUEUE_LABEL, TEST_QUEUE_NAME, TEST_WORKER_NAME
+from .literals import (
+    TEST_INTERVAL_SCHEDULE_EVERY, TEST_INTERVAL_SCHEDULE_PERIOD,
+    TEST_PERIODIC_TASK_NAME, TEST_PERIODIC_TASK_TASK, TEST_QUEUE_LABEL,
+    TEST_QUEUE_NAME, TEST_WORKER_NAME
+)
+
+
+class TaskManagerManagementCommandTestMixin:
+    def setUp(self):
+        super().setUp()
+
+        self._test_interval_schedule = IntervalSchedule.objects.create(
+            every=TEST_INTERVAL_SCHEDULE_EVERY,
+            period=TEST_INTERVAL_SCHEDULE_PERIOD
+        )
+        PeriodicTask.objects.create(
+            interval=self._test_interval_schedule,
+            name=TEST_PERIODIC_TASK_NAME, task=TEST_PERIODIC_TASK_TASK
+        )
 
 
 class TaskManagerTestMixin: